% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid21610720.R
\name{number1}
\alias{number1}
\title{Number of times for dice number 1}
\usage{
number1(n)
}
\description{
When you throw a coin n times and throw the dice as many times as the front, the number of times 1 comes out.
}
\examples{

number1(100)
}
